requirements.txt
requests
influxdb-client
python-dotenv